+++ ui/packages/app/grafana-panel-plugin/src/ParcaPanel.tsx
@@ -48,6 +48,13 @@ export const ParcaPanel: React.FC<Props> = ({ data, width, height }) => {
 
   const { flamegraphData, topTableData, actions } = response;
 
+  let total = BigInt(0);
+  if (flamegraphData.total !== undefined) {
+    total = BigInt(flamegraphData.total);
+  } else if (topTableData.total !== undefined) {
+    total = BigInt(topTableData.total);
+  }
+
   return (
     <Provider store={parcaStore}>
       <div
@@ -60,6 +67,8 @@ export const ParcaPanel: React.FC<Props> = ({ data, width, height }) => {
         )}
       >
         <ProfileView
+          total={total}
+          filtered={BigInt(0)}
           flamegraphData={flamegraphData}
           topTableData={topTableData}
           sampleUnit={flamegraphData.data?.unit ?? 'bytes'}