+++ ui/packages/app/grafana-datasource-plugin/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.113 (2023-04-04)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.112 (2023-04-01)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.111 (2023-03-31)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## [0.0.110](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.109...@parca/grafana-datasource@0.0.110) (2023-03-29)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.109 (2023-03-29)
+
+# 0.17.0 (2023-03-29)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
 ## [0.0.108](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.107...@parca/grafana-datasource@0.0.108) (2023-03-28)
 
 **Note:** Version bump only for package @parca/grafana-datasource